DIDBlockchain/requirements.txt
# DIDBlockchain\requirements.txt
asgiref==3.8.1
asttokens==2.4.1  # New
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6  # New
comm==0.2.2  # New
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.3
debugpy==1.8.7  # New
decorator==5.1.1  # New
Django==5.1.2
django-cors-headers==4.5.0
django-encrypted-model-fields==0.6.5
django-extensions==3.2.3
djangorestframework==3.15.2
executing==2.1.0  # New
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.29.5  # New
ipython==8.29.0  # New
itypes==1.2.0
jedi==0.19.1  # New
Jinja2==3.1.4
jupyter_client==8.6.3  # New
jupyter_core==5.7.2  # New
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0  # New
packaging==24.1
parso==0.8.4  # New
pip==24.2
platformdirs==4.3.6  # New
pluggy==1.5.0
prompt_toolkit==3.0.48
psutil==6.1.0  # New
psycopg2-binary==2.9.10
pure_eval==0.2.3  # New
pycparser==2.22
Pygments==2.18.0  # New
pytest==8.3.3
python-dateutil==2.9.0.post0
python-decouple==3.8
# pywin32==308  # New
PyYAML==6.0.2
pyzmq==26.2.0  # New
requests==2.32.3
setuptools==75.2.0
six==1.16.0  # New
sqlparse==0.5.1
stack-data==0.6.3  # New
tornado==6.4.1
traitlets==5.14.3  # New
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13  # New
